Add plumbing to build and package and integrate the new matlab generator.
So "npm run build" can build it. That is, once there's something
worth building...

Creates the matlab_compressed.js and matlab_compressed.js.map
files which implement the language generators. Note these are
stored in the build/ subfolder, but for use with demos etc., we
need to copy them to the blockly/ root folder.

The full command for rebuilding and automatic copy of files
into proper places is:

npm run build ; npm run checkin:built

Signed-off-by: Mario Kleiner <mario.kleiner.de@gmail.com>